# swagger/swagger.yaml
openapi: 3.0.3
info:
  description: "A.P.I de Produtos"
  version: "1.0.0"
  title: "A.P.I de Listagem de Produtos"
servers:
  - url: "http://localhost:5000"
    description: Servidor local
tags:
  - name: "produto"
    description: "Operações sobre produtos de Super-mercados"
paths:
  /produto:
    get:
      tags:
        - "produto"
      summary: "Obter lista os produtos de Super-mercados"
      responses:
        200:
          description: "Lista de Produto"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Produto"
    post:
      tags:
        - "produto"
      summary: "Cadastrar um novo produto"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Produto"
      responses:
        201:
          description: "produtos criado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produto"
  /produto/{produtoId}:
    put:
      tags:
        - "produto"
      summary: "Atualizar um produto já existente"
      parameters:
        - name: "produtoId"
          in: "path"
          required: true
          schema:
            type: "integer"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Produto"
      responses:
        200:
          description: "produto atualizado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produto"
        404:
          description: "produto não identificado"
    delete:
      tags:
        - "produto"
      summary: "Deletar um produto identificado"
      parameters:
        - name: "produtoId"
          in: "path"
          required: true
          schema:
            type: "integer"
      responses:
        204:
          description: "produto deletado"
        404:
          description: "produto não identificado"
components:
  schemas:
    Produto:
      type: "object"
      required:
        - "id"
        - "name"
        - "description"
      properties:
        id:
          type: "integer"
        name:
          type: "string"
        description:
          type: "string"